+++ Boa.Constrictor.Playwright/CHANGELOG.md
@@ -9,10 +9,12 @@ Its format is based on Keep a Changelog, and this project adheres to Semantic Ve
 ### Added
 
 - Added project `Boa.Constrictor.Playwright` for items that improve [Playwright](https://playwright.dev/dotnet/) support
-- Added `BrowseTheWebSynchronously` ability to manage playwright, browsers, and pages
+- Added `BrowseTheWebWithPlaywright` ability to manage playwright, browsers, and pages
 - Added `PlaywrightLocator` to store description + selector logic of a playwright locator
-- Added `AbstractPageTask` that makes it easier to write Tasks that use the BrowseTheWebSynchronously Ability
+- Added `AbstractPageTask` that makes it easier to write Tasks that use the BrowseTheWebWithPlaywright Ability
 - Added `AbstractLocatorTask` to make it easier to write tasks that perform operations on a single locator
+- Added `AbstractPageQuestion` to make it easier to write questions that use the BrowseTheWebWithPlaywright Ability
+- Added `AbstractLocatorQuestion` to make it easier to write questions that perform operations on a single locator
 - Added the following Tasks:
   - `OpenNewPage`
   - `Click`
@@ -25,6 +27,11 @@ Its format is based on Keep a Changelog, and this project adheres to Semantic Ve
   - `Clear`
   - `Focus`
   - `Press`
-- Added `Text` and `Attribute` questions
+- Added the following questions:
+  - `Text`
+  - `Attribute`
+  - `Title`
+  - `Visibility`
+  - `Enabled`
 - Added an `Expects` extension method to `IActor` to make it easier to use playwrights built in `ILocatorAssertions`
   - e.g. `await MyActor.Expects(WikiPage.Title).ToHaveTextAsync("Giant panda");`